+++ lib/ts/recipe/session/types.ts
@@ -201,7 +201,7 @@ export type RecipeInterface = {
     }): Promise<SessionClaimValidator[]> | SessionClaimValidator[];
 
     getSession(input: {
-        accessToken: string;
+        accessToken: string | undefined;
         antiCsrfToken?: string;
         options?: VerifySessionOptions;
         userContext: any;