+++ CHANGELOG.md
@@ -7,6 +7,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [unreleased]
 
+## [14.0.2] - 2023-05-11
+
+### Changes
+
+-   Made the access token string optional in the overrideable `getSession` function
+-   Moved checking if the access token is defined into the overrideable `getSession` function
+
 ## [14.0.1] - 2023-05-11
 
 -   Fixes an issue where API key based login with dashboard would return invalid API key even if the entered API key was valid